+++ combust/CHANGELOG.md
@@ -5,9 +5,9 @@ The following tools were used to generate this project:
 
 The following steps were used to generate this project:
 - Create Angular project with ng: `ng new combust --defaults --skip-install --skip-git --no-standalone `.
-- Update angular.json with port.
 - Create project file (`combust.esproj`).
 - Create `launch.json` to enable debugging.
+- Create `nuget.config` to specify location of the JavaScript Project System SDK (which is used in the first line in `combust.esproj`).
 - Update package.json to add `jest-editor-support`.
 - Update `start` script in `package.json` to specify host.
 - Add `karma.conf.js` for unit tests.